Build the disclaimer message key with commentId
The test spelled out d.String() + "-comment", which is what commentId
does and what GetComment calls, so the convention lived in two places and
a change to one would have left the other passing on the wrong key.

commentId is unexported, so the two tests that only inspect disclaimer
metadata move into the package, alongside the internal tests already
there. The behavioural tests stay outside it, where they use the helpers
in helpers_test.go.
