+++ checker/disclaimer_test.go
@@ -2,54 +2,15 @@ package checker_test
 
 import (
 	"encoding/json"
-	"regexp"
 	"testing"
 
 	"github.com/getkin/kin-openapi/openapi3"
 	"github.com/oasdiff/oasdiff/checker"
-	"github.com/oasdiff/oasdiff/checker/localizations"
 	"github.com/oasdiff/oasdiff/diff"
 	"github.com/oasdiff/oasdiff/formatters"
 	"github.com/oasdiff/oasdiff/load"
 	"github.com/stretchr/testify/require"
 )
-
-func allDisclaimers(t *testing.T) []checker.Disclaimer {
-	t.Helper()
-
-	unnamed := checker.Disclaimer(0).String()
-	var all []checker.Disclaimer
-	for d := checker.Disclaimer(1); d.String() != unnamed; d++ {
-		all = append(all, d)
-	}
-	require.NotEmpty(t, all)
-	return all
-}
-
-// Names are a public surface: they appear in output and are the key a caller
-// would use to configure a disclaimer. They have to be unique and stable.
-func TestDisclaimerNames(t *testing.T) {
-	kebab := regexp.MustCompile(`^[a-z0-9]+(-[a-z0-9]+)*$`)
-	seen := map[string]bool{}
-	for _, d := range allDisclaimers(t) {
-		name := d.String()
-		require.Truef(t, kebab.MatchString(name), "disclaimer name %q is not kebab-case", name)
-		require.Falsef(t, seen[name], "duplicate disclaimer name %q", name)
-		seen[name] = true
-	}
-}
-
-// Every disclaimer explains itself, in every language, or it reaches a user as
-// a bare identifier.
-func TestDisclaimersAreLocalized(t *testing.T) {
-	for _, lang := range localizations.GetSupportedLanguages() {
-		localizer := checker.NewLocalizer(lang)
-		for _, d := range allDisclaimers(t) {
-			key := d.String() + "-comment"
-			require.NotEqualf(t, key, localizer(key), "disclaimer %q has no %s text", d, lang)
-		}
-	}
-}
 
 // A change compared branch by branch cannot be reported at error, whatever the
 // rule declares, because a sibling branch may still guarantee what this one
